bot/requirements.txt
aiogram>=3.7,<3.8
asyncpg~=0.29
httpx~=0.27
timezonefinder~=6.4
pyswisseph==2.10.3.2
jsonschema~=4.23
tzdata
lunar-python==1.4.4
openai>=1.30,<2
jsonschema>=4.22